+++ app/src/main/java/org/mozilla/fenix/library/bookmarks/BookmarkComponent.kt
@@ -13,7 +13,9 @@ import org.mozilla.fenix.mvi.Reducer
 import org.mozilla.fenix.mvi.UIComponent
 import org.mozilla.fenix.mvi.UIView
 import org.mozilla.fenix.mvi.ViewState
+import org.mozilla.fenix.test.Mockable
 
+@Mockable
 class BookmarkComponent(
     private val container: ViewGroup,
     bus: ActionBusFactory,
@@ -28,7 +30,28 @@ class BookmarkComponent(
     override val reducer: Reducer<BookmarkState, BookmarkChange> = { state, change ->
         when (change) {
             is BookmarkChange.Change -> {
-                state.copy(tree = change.tree)
+                val mode =
+                    if (state.mode is BookmarkState.Mode.Selecting) {
+                        BookmarkState.Mode.Selecting(state.mode.selectedItems.filter {
+                            it in change.tree
+                        }.toSet())
+                    } else state.mode
+                state.copy(tree = change.tree, mode = mode)
+            }
+            is BookmarkChange.IsSelected -> {
+                val selectedItems = if (state.mode is BookmarkState.Mode.Selecting) {
+                    state.mode.selectedItems + change.newlySelectedItem
+                } else setOf(change.newlySelectedItem)
+                state.copy(mode = BookmarkState.Mode.Selecting(selectedItems))
+            }
+            is BookmarkChange.IsDeselected -> {
+                val selectedItems = if (state.mode is BookmarkState.Mode.Selecting) {
+                    state.mode.selectedItems - change.newlyDeselectedItem
+                } else setOf()
+                val mode = if (selectedItems.isEmpty()) BookmarkState.Mode.Normal else BookmarkState.Mode.Selecting(
+                    selectedItems
+                )
+                state.copy(mode = mode)
             }
         }
     }
@@ -44,7 +67,7 @@ class BookmarkComponent(
 data class BookmarkState(val tree: BookmarkNode?, val mode: BookmarkState.Mode) : ViewState {
     sealed class Mode {
         object Normal : Mode()
-        data class Selecting(val selectedItems: List<BookmarkNode>) : Mode()
+        data class Selecting(val selectedItems: Set<BookmarkNode>) : Mode()
     }
 }
 
@@ -57,11 +80,18 @@ sealed class BookmarkAction : Action {
     data class OpenInNewTab(val item: BookmarkNode) : BookmarkAction()
     data class OpenInPrivateTab(val item: BookmarkNode) : BookmarkAction()
     data class Select(val item: BookmarkNode) : BookmarkAction()
+    data class Deselect(val item: BookmarkNode) : BookmarkAction()
     data class Delete(val item: BookmarkNode) : BookmarkAction()
-    object ExitSelectMode : BookmarkAction()
     object BackPressed : BookmarkAction()
+    object ModeChanged : BookmarkAction()
 }
 
 sealed class BookmarkChange : Change {
     data class Change(val tree: BookmarkNode) : BookmarkChange()
+    data class IsSelected(val newlySelectedItem: BookmarkNode) : BookmarkChange()
+    data class IsDeselected(val newlyDeselectedItem: BookmarkNode) : BookmarkChange()
+}
+
+operator fun BookmarkNode.contains(item: BookmarkNode): Boolean {
+    return children?.contains(item) ?: false
 }