+++ app/src/main/java/org/mozilla/fenix/library/bookmarks/BookmarkFragment.kt
@@ -4,6 +4,12 @@
 
 package org.mozilla.fenix.library.bookmarks
 
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
+import android.graphics.PorterDuff
+import android.graphics.PorterDuffColorFilter
+import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.Menu
@@ -11,7 +17,12 @@ import android.view.MenuInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.view.menu.ActionMenuItemView
+import androidx.appcompat.widget.ActionMenuView
+import androidx.appcompat.widget.Toolbar
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
 import kotlinx.android.synthetic.main.fragment_bookmark.view.*
@@ -20,6 +31,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.Dispatchers.Main
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import mozilla.appservices.places.BookmarkRoot
 import mozilla.components.concept.storage.BookmarkNode
@@ -32,6 +44,8 @@ import org.mozilla.fenix.BrowserDirection
 import org.mozilla.fenix.BrowsingModeManager
 import org.mozilla.fenix.HomeActivity
 import org.mozilla.fenix.R
+import org.mozilla.fenix.components.FenixSnackbar
+import org.mozilla.fenix.ext.getColorFromAttr
 import org.mozilla.fenix.ext.requireComponents
 import org.mozilla.fenix.ext.share
 import org.mozilla.fenix.mvi.ActionBusFactory
@@ -46,7 +60,7 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
     private lateinit var job: Job
     private lateinit var bookmarkComponent: BookmarkComponent
     private lateinit var signInComponent: SignInComponent
-    private lateinit var currentRoot: BookmarkNode
+    private var currentRoot: BookmarkNode? = null
 
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Main + job
@@ -61,12 +75,12 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         job = Job()
+        (activity as AppCompatActivity).title = getString(R.string.library_bookmarks)
         setHasOptionsMenu(true)
     }
 
     override fun onResume() {
         super.onResume()
-        (activity as AppCompatActivity).title = getString(R.string.library_bookmarks)
         (activity as AppCompatActivity).supportActionBar?.show()
         checkIfSignedIn()
     }
@@ -84,7 +98,67 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.library_menu, menu)
+        val toolbar = (activity as AppCompatActivity).findViewById<Toolbar>(R.id.navigationToolbar)
+        when (val mode = (bookmarkComponent.uiView as BookmarkUIView).mode) {
+            BookmarkState.Mode.Normal -> {
+                inflater.inflate(R.menu.library_menu, menu)
+                activity?.title =
+                    if (currentRoot?.title in setOf(
+                            "root",
+                            null
+                        )
+                    ) getString(R.string.library_bookmarks) else currentRoot!!.title
+                toolbar.setBackgroundColor(R.attr.toolbarColor.getColorFromAttr(context!!))
+                toolbar.setTitleTextColor(R.attr.toolbarTextColor.getColorFromAttr(context!!))
+            }
+            is BookmarkState.Mode.Selecting -> {
+                inflater.inflate(R.menu.bookmarks_select_multi, menu)
+                activity?.title = getString(R.string.bookmarks_multi_select_title, mode.selectedItems.size)
+                val colorFilter =
+                    PorterDuffColorFilter(ContextCompat.getColor(context!!, R.color.off_white), PorterDuff.Mode.SRC_IN)
+                themeToolbar(toolbar, colorFilter)
+            }
+        }
+    }
+
+    private fun themeToolbar(
+        toolbar: Toolbar,
+        colorFilter: PorterDuffColorFilter
+    ) {
+        toolbar.setBackgroundColor(ContextCompat.getColor(context!!, R.color.bookmark_multi_select_actionbar))
+        toolbar.setTitleTextColor(ContextCompat.getColor(context!!, R.color.off_white))
+        toolbar.overflowIcon?.colorFilter = colorFilter
+        (0 until toolbar.childCount).forEach {
+            when (val item = toolbar.getChildAt(it)) {
+                is ImageButton -> item.drawable.colorFilter = colorFilter
+                is ActionMenuView -> themeActionMenuView(item, colorFilter)
+            }
+        }
+    }
+
+    private fun themeActionMenuView(
+        item: ActionMenuView,
+        colorFilter: PorterDuffColorFilter
+    ) {
+        (0 until item.childCount).forEach {
+            val innerChild = item.getChildAt(it)
+            if (innerChild is ActionMenuItemView) {
+                themeChildren(innerChild, item, colorFilter)
+            }
+        }
+    }
+
+    private fun themeChildren(
+        innerChild: ActionMenuItemView,
+        item: ActionMenuView,
+        colorFilter: PorterDuffColorFilter
+    ) {
+        val drawables = innerChild.compoundDrawables
+        for (k in drawables.indices) {
+            drawables[k]?.let {
+                item.post { innerChild.compoundDrawables[k].colorFilter = colorFilter }
+            }
+        }
     }
 
     @SuppressWarnings("ComplexMethod")
@@ -116,10 +190,15 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
                             )
                     }
                     is BookmarkAction.Select -> {
-                        ItsNotBrokenSnack(context!!).showSnackbar(issueNumber = "1239")
+                        getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.IsSelected(it.item))
+                    }
+                    is BookmarkAction.Deselect -> {
+                        getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.IsDeselected(it.item))
                     }
                     is BookmarkAction.Copy -> {
-                        ItsNotBrokenSnack(context!!).showSnackbar(issueNumber = "1239")
+                        it.item.copyUrl(context!!)
+                        FenixSnackbar.make(view!!, FenixSnackbar.LENGTH_LONG)
+                            .setText(context!!.getString(R.string.url_copied)).show()
                     }
                     is BookmarkAction.Share -> {
                         it.item.url?.let { url -> requireContext().share(url) }
@@ -139,15 +218,10 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
                     is BookmarkAction.Delete -> {
                         launch(IO) {
                             requireComponents.core.bookmarksStorage.deleteNode(it.item.guid)
-                            requireComponents.core.bookmarksStorage.getTree(currentRoot.guid, false)
-                                ?.let { node ->
-                                    getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.Change(node))
-                                }
+                            refreshBookmarks()
                         }
                     }
-                    is BookmarkAction.ExitSelectMode -> {
-                        ItsNotBrokenSnack(context!!).showSnackbar(issueNumber = "1239")
-                    }
+                    is BookmarkAction.ModeChanged -> activity?.invalidateOptionsMenu()
                 }
             }
 
@@ -162,6 +236,7 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
             }
     }
 
+    @SuppressWarnings("ComplexMethod")
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
             R.id.libraryClose -> {
@@ -171,6 +246,47 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
             }
             R.id.librarySearch -> {
                 // TODO Library Search
+                true
+            }
+            R.id.open_bookmarks_in_new_tabs_multi_select -> {
+                getSelectedBookmarks().forEach { node ->
+                    node.url?.let {
+                        requireComponents.useCases.tabsUseCases.addTab.invoke(it)
+                    }
+                }
+
+                (activity as HomeActivity).browsingModeManager.mode = BrowsingModeManager.Mode.Normal
+                Navigation.findNavController(requireActivity(), R.id.container)
+                    .navigate(BookmarkFragmentDirections.actionBookmarkFragmentToHomeFragment())
+                true
+            }
+            R.id.share_bookmarks_multi_select -> {
+                ItsNotBrokenSnack(context!!).showSnackbar(issueNumber = "1539")
+                true
+            }
+            R.id.open_bookmarks_in_private_tabs_multi_select -> {
+                getSelectedBookmarks().forEach { node ->
+                    node.url?.let {
+                        requireComponents.useCases.tabsUseCases.addPrivateTab.invoke(it)
+                    }
+                }
+
+                (activity as HomeActivity).browsingModeManager.mode = BrowsingModeManager.Mode.Private
+                Navigation.findNavController(requireActivity(), R.id.container)
+                    .navigate(BookmarkFragmentDirections.actionBookmarkFragmentToHomeFragment())
+                true
+            }
+            R.id.delete_bookmarks_multi_select -> {
+                val deleteJob = launch(IO) {
+                    delay(bookmarkDeletionDelay)
+                    deleteSelectedBookmarks()
+                    refreshBookmarks()
+                }
+                FenixSnackbar.make(view!!, FenixSnackbar.LENGTH_LONG)
+                    .setText(getString(R.string.bookmark_deletion_multiple_snackbar_message))
+                    .setAction(getString(R.string.bookmark_undo_deletion)) {
+                        deleteJob.cancel()
+                    }.show()
                 true
             }
             else -> super.onOptionsItemSelected(item)
@@ -186,8 +302,8 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
             currentRoot = requireComponents.core.bookmarksStorage.getTree(currentGuid) as BookmarkNode
 
             launch(Main) {
-                if (currentGuid != BookmarkRoot.Root.id) (activity as HomeActivity).title = currentRoot.title
-                getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.Change(currentRoot))
+                if (currentGuid != BookmarkRoot.Root.id) (activity as HomeActivity).title = currentRoot!!.title
+                getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.Change(currentRoot!!))
             }
         }
     }
@@ -206,5 +322,30 @@ class BookmarkFragment : Fragment(), CoroutineScope, BackHandler, AccountObserve
     }
 
     override fun onProfileUpdated(profile: Profile) {
+    }
+
+    private fun getSelectedBookmarks() = (bookmarkComponent.uiView as BookmarkUIView).getSelected()
+
+    private suspend fun deleteSelectedBookmarks() {
+        getSelectedBookmarks().forEach {
+            requireComponents.core.bookmarksStorage.deleteNode(it.guid)
+        }
+    }
+
+    private suspend fun refreshBookmarks() {
+        requireComponents.core.bookmarksStorage.getTree(currentRoot!!.guid, false)
+            ?.let { node ->
+                getManagedEmitter<BookmarkChange>().onNext(BookmarkChange.Change(node))
+            }
+    }
+
+    private fun BookmarkNode.copyUrl(context: Context) {
+        val clipBoard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+        val uri = Uri.parse(url)
+        clipBoard.primaryClip = ClipData.newRawUri("Uri", uri)
+    }
+
+    companion object {
+        private const val bookmarkDeletionDelay = 3000L
     }
 }